+++ common/src/test/java/io/piano/android/common/UserAgentInterceptorTest.kt
@@ -1,4 +1,4 @@
-package io.piano.android.id
+package io.piano.android.common
 
 import okhttp3.OkHttpClient
 import okhttp3.Request