+++ composer/src/main/java/io/piano/android/composer/model/EventExecutionContext.kt
@@ -15,4 +15,5 @@ public class EventExecutionContext(
     @JvmField public val userSegments: UserSegmentsContainer,
     @JvmField public val accessList: List<Access>?,
     @JvmField public val activeMeters: List<ActiveMeter>?,
+    @JvmField public val sessions: List<SessionStats>?,
 )